Trying landscape lock again in Game-scoring
Read this:
"The problem is not with any plugin. The reason for this not working is that Cordova doesn't run with the command ionic serve. So it'll not work in the inspect tab. Use an emulator instead, it'll work just fine." Hoping that after a fork into my branch it will work.